requirements-test.txt
# 测试依赖包
# 用于VoiceHelper项目的各种测试

# 核心测试框架
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.0.0  # 并行测试
pytest-html>=3.1.0   # HTML报告

# HTTP客户端和服务器测试
requests>=2.28.0
aiohttp>=3.8.0
httpx>=0.24.0

# WebSocket测试
websockets>=11.0.0

# 性能和负载测试
locust>=2.14.0
psutil>=5.9.0

# Web UI测试
selenium>=4.8.0
webdriver-manager>=3.8.0

# 数据处理和验证
pydantic>=1.10.0
jsonschema>=4.17.0

# 模拟和存根
responses>=0.23.0
factory-boy>=3.2.0
faker>=18.0.0

# 数据库测试
pytest-postgresql>=4.1.0
pytest-redis>=3.0.0

# 时间和日期处理
freezegun>=1.2.0

# 日志和调试
pytest-logging>=2016.11.4
pytest-benchmark>=4.0.0

# 代码质量
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# 安全测试
bandit>=1.7.0
safety>=2.3.0

# 文档测试
pytest-doctestplus>=0.12.0

# 环境管理
python-dotenv>=1.0.0

# 数据生成
mimesis>=7.0.0

# API测试
tavern>=2.0.0

# 性能分析
py-spy>=0.3.0
memory-profiler>=0.60.0

# 并发测试
pytest-parallel>=0.1.0

# 测试数据
pytest-datadir>=1.4.0

# 网络测试
pytest-httpserver>=1.0.0

# 容器测试
testcontainers>=3.7.0

# 云服务测试
moto>=4.1.0  # AWS服务模拟

# 时间序列测试
pytest-timeout>=2.1.0

# 参数化测试增强
pytest-cases>=3.6.0

# 测试报告增强
allure-pytest>=2.12.0

# 性能回归测试
pytest-benchmark>=4.0.0

# 内存泄漏检测
pympler>=0.9

# 代码覆盖率增强
coverage[toml]>=7.0.0

# API文档测试
schemathesis>=3.17.0

# 数据库迁移测试
alembic>=1.10.0

# 消息队列测试
kombu>=5.2.0

# 缓存测试
redis>=4.5.0

# 文件系统测试
pyfakefs>=5.1.0

# 网络延迟模拟
netifaces>=0.11.0

# 加密测试
cryptography>=40.0.0

# 配置测试
dynaconf>=3.1.0

# 监控测试
prometheus-client>=0.16.0

# 日志测试
structlog>=23.1.0

# 异步测试增强
pytest-trio>=0.8.0
pytest-tornado>=0.8.0

# 数据验证测试
cerberus>=1.3.0
marshmallow>=3.19.0

# 图像处理测试（如果需要）
Pillow>=9.5.0

# 音频处理测试（如果需要）
# librosa>=0.10.0  # 取消注释如果需要音频测试

# 机器学习测试（如果需要）
# scikit-learn>=1.2.0  # 取消注释如果需要ML测试
# numpy>=1.24.0
# pandas>=2.0.0

# 自然语言处理测试（如果需要）
# transformers>=4.28.0  # 取消注释如果需要NLP测试
# torch>=2.0.0

# 开发工具
ipython>=8.12.0
jupyter>=1.0.0
